+++ api/python/quilt3/admin/buckets.py
@@ -1,6 +1,7 @@
 import typing as T
 
-from . import _graphql_client, exceptions, types, util
+from .. import _graphql_client
+from . import exceptions, types, util
 
 
 def get(name: str) -> T.Optional[types.Bucket]: